services/image_analyzer.py
import base64
from pathlib import Path
from typing import Optional, Dict
from PIL import Image
import io


class ImageAnalyzer:
    """Сервис для анализа фото с едой"""
    
    def __init__(self):
        self.cache_dir = Path("cache")
        self.cache_dir.mkdir(exist_ok=True)
    
    async def analyze_food_image(self, image_path: str) -> Optional[str]:
        """
        Анализирует изображение еды и возвращает описание продуктов
        
        Args:
            image_path: Путь к изображению
            
        Returns:
            Текстовое описание продуктов на изображении
        """
        try:
            # Открываем и проверяем изображение
            image = Image.open(image_path)
            
            # В реальном проекте здесь можно использовать:
            # 1. OpenAI Vision API для распознавания еды
            # 2. Google Cloud Vision API
            # 3. Специализированные модели для распознавания еды
            
            # Для демонстрации возвращаем инструкцию
            return self._get_simple_analysis(image)
            
        except Exception as e:
            print(f"Ошибка при анализе изображения: {e}")
            return None
    
    def _get_simple_analysis(self, image: Image.Image) -> str:
        """
        Простой анализ изображения (заглушка)
        В реальном проекте здесь должно быть использование AI API
        """
        # Базовая информация об изображении
        width, height = image.size
        
        return (
            "Для точного анализа фото требуется интеграция с AI моделью.\n"
            "Пожалуйста, опишите ваше блюдо текстом:\n"
            "Например: '200g куриная грудка, 100g рис, салат'"
        )
    
    async def download_and_save(self, file_path: str, file_id: str) -> Path:
        """
        Сохраняет загруженное изображение
        
        Args:
            file_path: Путь к файлу от Telegram
            file_id: ID файла
            
        Returns:
            Путь к сохраненному файлу
        """
        save_path = self.cache_dir / f"{file_id}.jpg"
        return save_path

